requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
attrs==16.3.0
autobahn==0.18.2
Automat==0.5.0
backports.ssl-match-hostname==3.5.0.1
cffi==1.10.0
constantly==15.1.0
cryptography==1.8.1
decorator==4.0.11
docutils==0.13.1
enum34==1.1.6
freeze==1.0.10
GDAX==0.2.2
idna==2.5
incremental==16.10.1
ipaddress==1.0.18
networkx==1.11
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
redis==2.10.5
requests==2.13.0
service-identity==16.0.0
six==1.10.0
statistics==1.0.3.5
Twisted==17.1.0
txaio==2.7.0
websocket-client==0.40.0
zope.interface==4.3.3
